refactor: `TableGroup`, `OrderTable` 의존성 제거
