+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,16 +1,15 @@
 package kitchenpos.table.domain;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import kitchenpos.table.exception.NotEnoughGuestsException;
 import kitchenpos.table.exception.OrderTableEmptyException;
 import kitchenpos.table.exception.OrderTableNotEmptyException;
-import kitchenpos.table.exception.TableGroupExistsException;
+import kitchenpos.tablegroup.exception.TableGroupExistsException;
 
 @Table(name = "order_table")
 @Entity
@@ -22,9 +21,8 @@ public class OrderTable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
-    @JoinColumn(name = "table_group_id")
-    private TableGroup tableGroup;
+    @Column(name = "table_group_id")
+    private Long tableGroupId;
 
     private int numberOfGuests;
 
@@ -33,14 +31,14 @@ public class OrderTable {
     protected OrderTable() {
     }
 
-    public OrderTable(TableGroup tableGroup, int numberOfGuests, boolean empty) {
-        this.tableGroup = tableGroup;
+    public OrderTable(Long tableGroupId, int numberOfGuests, boolean empty) {
+        this.tableGroupId = tableGroupId;
         this.numberOfGuests = numberOfGuests;
         this.empty = empty;
     }
 
-    public void changeTableGroup(TableGroup tableGroup) {
-        this.tableGroup = tableGroup;
+    public void changeTableGroup(Long tableGroupId) {
+        this.tableGroupId = tableGroupId;
     }
 
     public void changeEmpty(boolean empty) {
@@ -49,7 +47,7 @@ public class OrderTable {
     }
 
     public void validateTableGroupNotExists() {
-        if (tableGroup != null) {
+        if (tableGroupId != null) {
             throw new TableGroupExistsException();
         }
     }
@@ -86,8 +84,8 @@ public class OrderTable {
         return id;
     }
 
-    public TableGroup getTableGroup() {
-        return tableGroup;
+    public Long getTableGroupId() {
+        return tableGroupId;
     }
 
     public int getNumberOfGuests() {
