+++ src/main/java/kitchenpos/table/dto/response/OrderTableResponse.java
@@ -17,11 +17,11 @@ public class OrderTableResponse {
     }
 
     public static OrderTableResponse from(OrderTable orderTable) {
-        if (orderTable.getTableGroup() == null) {
+        if (orderTable.getTableGroupId() == null) {
             return new OrderTableResponse(orderTable.getId(), null,
                     orderTable.getNumberOfGuests(), orderTable.isEmpty());
         }
-        return new OrderTableResponse(orderTable.getId(), orderTable.getTableGroup().getId(),
+        return new OrderTableResponse(orderTable.getId(), orderTable.getTableGroupId(),
                 orderTable.getNumberOfGuests(), orderTable.isEmpty());
     }
