+++ src/main/java/kitchenpos/tablegroup/dto/request/CreateTableGroupRequest.java
@@ -1,8 +1,9 @@
-package kitchenpos.table.dto.request;
+package kitchenpos.tablegroup.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.sun.istack.NotNull;
 import java.util.List;
+import kitchenpos.table.dto.request.OrderTableRequest;
 
 public class CreateTableGroupRequest {
