+++ src/main/java/kitchenpos/tablegroup/exception/OrderTableCountNotEnoughException.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.exception;
+package kitchenpos.tablegroup.exception;
 
 public class OrderTableCountNotEnoughException extends RuntimeException {
