+++ src/main/java/kitchenpos/tablegroup/exception/TableGroupExistsException.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.exception;
+package kitchenpos.tablegroup.exception;
 
 public class TableGroupExistsException extends RuntimeException {
