+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import java.time.LocalDateTime;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class TableGroupFixture {
