+++ src/test/java/kitchenpos/supports/ServiceTestContext.java
@@ -11,8 +11,8 @@ import kitchenpos.order.service.OrderService;
 import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.product.service.ProductService;
 import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.table.repository.TableGroupRepository;
-import kitchenpos.table.service.TableGroupService;
+import kitchenpos.tablegroup.repository.TableGroupRepository;
+import kitchenpos.tablegroup.service.TableGroupService;
 import kitchenpos.table.service.TableService;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
